requirements.txt
asgiref
Django
crispy-bootstrap5
django-jalali
humanize
jdatetime
Pillow
pytz
sqlparse
